requirements.txt
gradio>=3.50
ollama>=0.1.6
python-dotenv>=0.19.0